Clean up, clean up!
TLDR: Minor adjustments, white boxes, and curves.

Adjusted the white boxes to properly fit. Added in the new curve types. Added in some of the detailing on a few of the white boxes. Made silhouette effect on Employee Icons  and animation, though may need to adjust later. Fixed the video scroll height problem. Made the button on home work. Minor effects here and there that we might scrap. Will need to adjust the text boxes and the arrows for the employee boxes. Basically finished 90% of the detailing besides the bottom of Home and Projects.