Project Page Debug
Made the title follow the sliders when the F/O section is activated. Made the white path resize properly. Made the sprite follow correctly. All that's left for this section is the resizing problem that we'll have to apply the principles of to all of the pages. This being in how the text slips out of the white box when the sliders are crushed too much.